requirements.txt
langchain==0.1.14
langchain-openai==0.1.9
langchain-community==0.0.28
python-dotenv==1.0.0
openai==1.14.1

